Issue 1 Resolved : Memory errors resolved : Code Refactoring resolved
