+++ Makefile
@@ -8,17 +8,19 @@ pathfind:
 
 	mkdir -p assets
 	$(CC) $(CFLAGS) -c -o assets/List.o src/List/List.c
-	$(CC) $(CFLAGS) -c -o assets/pathfind.o src/pathfind.c
 	$(CC) $(CFLAGS) -c -o assets/main.o src/main.c
-	$(CC) $(CFLAGS) -o $(OUT) assets/main.o assets/pathfind.o assets/List.o -lm
+	$(CC) $(CFLAGS) -c -o assets/pathfind.o src/pathfind.c
+	$(CC) $(CFLAGS) -c -o assets/board.o src/board.c
+	$(CC) $(CFLAGS) -o $(OUT) assets/main.o assets/pathfind.o assets/board.o assets/List.o -lm
 
 debug:
 
 	mkdir -p assets
 	$(CC) $(CFLAGSD) -c -o assets/List.o src/List/List.c
-	$(CC) $(CFLAGSD) -c -o assets/pathfind.o src/pathfind.c
 	$(CC) $(CFLAGSD) -c -o assets/main.o src/main.c
-	$(CC) $(CFLAGSD) -o $(OUT) assets/main.o assets/pathfind.o assets/List.o -lm
+	$(CC) $(CFLAGSD) -c -o assets/pathfind.o src/pathfind.c
+	$(CC) $(CFLAGSD) -c -o assets/board.o src/board.c
+	$(CC) $(CFLAGSD) -o $(OUT) assets/main.o assets/pathfind.o assets/board.o assets/List.o -lm
 
 clean:
