--- issue1/src/Boards/Invalid/invalid1.txt
@@ -1,3 +1,0 @@
-#S#
-# #
-#E#
