--- issue1/src/Boards/Invalid/invalid2.txt
@@ -1,3 +1,0 @@
-#S##
-#_#
-#E#
