--- issue1/src/Boards/Invalid/invalid3.txt
@@ -1,4 +1,0 @@
-#S#
-#_#
-#E#
-#E#
