--- issue1/src/Boards/test1.txt
@@ -1,3 +1,0 @@
-#S#
-#_#
-#E#
