--- issue1/src/Boards/test2.txt
@@ -1,5 +1,0 @@
-#s#####
-#___###
-#_#_###
-#_____e
-#######
