--- issue1/src/Boards/test3.txt
@@ -1,5 +1,0 @@
-S########_##
-_###_###_#E#
-__##_###_###
-___________#
-############
