--- issue1/src/Boards/test4.txt
@@ -1,8 +1,0 @@
-####
-##S#
-##_#
-##_#
-E#_#
-##_#
-##_#
-####
