--- issue1/src/Boards/test5.txt
@@ -1,16 +1,0 @@
-###________________E#########
-#___###################___#_#
-#_####________________#_#_#_#
-#_###################_#_#_#_#
-#_____________________#_#_#_#
-#####################_#_#_#_#
-#___##________________#_#_#_#
-#_#_##_###_##_#########_#_#_#
-#_#____#___##_#_________#_#_#
-#_#_##_################_#_#_#
-###_##_____________####_#_#_#
-###_##############_##_#_#_#_#
-###_____________##____#_#_#_#
-######_########_#######_#_#_#
-######_####_____________#___#
-S______######################
